refactor: migrate ingest/indexer callers off object_store shim
