+++ tests/e2e/test_panel_to_promotion_consume.py
@@ -13,7 +13,7 @@ from app.events.types import PROMOTE_DONE, PROMOTE_INTENT_CREATED
 from app.observability.status_service import get_system_status
 from app.promotion.consumer import consume_promotion_intents
 from app.store import object_store as object_store_module
-from app.store.object_store import DomainObject, ObjectStore
+from app.objects import DomainObject, ObjectStore
 
 pytestmark = pytest.mark.not_pg
